+++ system/lib/pthread/library_pthread.c
@@ -968,7 +968,7 @@ int EMSCRIPTEN_KEEPALIVE proxy_main(int argc, char** argv) {
       // Proceed by running main() on the main browser thread as a fallback.
       return __call_main(_main_arguments.argc, _main_arguments.argv);
     }
-    EM_ASM(Module['noExitRuntime'] = true);
+    EM_ASM(noExitRuntime = true);
     return 0;
   } else {
     return __call_main(_main_arguments.argc, _main_arguments.argv);