+++ tests/test_keyboard_codes.c
@@ -79,5 +79,5 @@ int main()
   emscripten_set_keydown_callback(EMSCRIPTEN_EVENT_TARGET_WINDOW, 0, 1, key_callback);
   emscripten_set_keyup_callback(EMSCRIPTEN_EVENT_TARGET_WINDOW, 0, 1, key_callback);
   emscripten_set_keypress_callback(EMSCRIPTEN_EVENT_TARGET_WINDOW, 0, 1, key_callback);
-  EM_ASM(Module['noExitRuntime'] = true);
+  EM_ASM(noExitRuntime = true);
 }